// src/layouts/layout.types.ts
type RequiredProps = {
    type: StatblockItemType;
    conditioned?: boolean;
    id: string;
};

export type CommonProps = RequiredProps & {
    properties: string[];
    fallback?: string;
    hasRule?: boolean;
    dice?: boolean;
    diceProperty?: keyof Participant;
    diceText?: string;
    diceCallback?: string;
    doNotAddClass?: boolean;
};

export const MarkdownTypes = [
    "property",
    "traits",
    "spells",
    "text",
    "saves"
] as const;

type GenericTextProp = {
    markdown?: boolean;
};

type GroupProps = {
    type: "group";
    heading?: string;
    headingProp?: boolean;
    nested: StatblockItem[];
    cls?: string;
};
type HeadingProps = {
    type: "heading";
    size: number;
};
type InlineProps = {
    type: "inline";
    heading?: string;
    headingProp?: boolean;
    nested: StatblockItem[];
    cls?: string;
};

type PropertyProps = {
    type: "property";
    callback?: string;
    display?: string;
};
type SavesProps = {
    type: "saves";
    callback?: string;
    display?: string;
};
type SpellsProps = {
    type: "spells";
    callback?: string;
    heading?: string;
};
type SubHeadingProps = {
    type: "subheading";
    separator?: string;
};
type TableProps = {
    type: "table";
    headers: string[];
    calculate: boolean;
    modifier?: string;
};
type TraitsProps = {
    type: "traits";
    callback?: string;
    heading?: string;
    headingProp?: boolean;
    subheadingText?: string;
};
type ImageProps = {
    type: "image";
    heading?: string;
};
type TextProps = {
    type: "text";
    heading?: string;
    headingProp?: boolean;
    text: string;
};
export type IfElseCondition = {
    condition: string;
    nested: [GroupItem];
};
type IfElseProps = {
    type: "ifelse";
    conditions: IfElseCondition[];
};
type CollapseProps = {
    type: "collapse";
    nested: [GroupItem];
    heading?: string;
    hasRule?: boolean;
    open: boolean;
};
type JavaScriptProps = {
    type: "javascript";
    code: string;
};
type LayoutProps = {
    type: "layout";
    layout: string;
};

type ActionProps = {
    type: "action";
    action?: string;
    callback?: string;
    icon?: string;
};

export type GroupItem = CommonProps & GroupProps;
export type HeadingItem = CommonProps & HeadingProps;
export type InlineItem = CommonProps & InlineProps;
export type PropertyItem = CommonProps & PropertyProps & GenericTextProp;
export type SavesItem = CommonProps & SavesProps & GenericTextProp;
export type TraitsItem = CommonProps & TraitsProps & GenericTextProp;
export type SpellsItem = CommonProps & SpellsProps & GenericTextProp;
export type SubHeadingItem = CommonProps & SubHeadingProps;
export type TableItem = CommonProps & TableProps;
export type ImageItem = CommonProps & ImageProps;
export type TextItem = CommonProps & TextProps & GenericTextProp;

export type ActionItem = RequiredProps & ActionProps;
export type IfElseItem = RequiredProps & IfElseProps;
export type CollapseItem = RequiredProps & CollapseProps;
export type JavaScriptItem = RequiredProps & JavaScriptProps;
export type LayoutItem = RequiredProps & LayoutProps;

export type StatblockItem =
    | GroupItem
    | HeadingItem
    | InlineItem
    | PropertyItem
    | SavesItem
    | TraitsItem
    | SpellsItem
    | SubHeadingItem
    | TableItem
    | ImageItem
    | TextItem
    | IfElseItem
    | CollapseItem
    | JavaScriptItem
    | LayoutItem
    | ActionItem;
// ...
export interface Layout {
    name: string;
    id: string;
    diceParsing?: DiceParsing[];
    cssProperties?: Partial<LayoutCSSProperties>;
    columns?: number;
    forceColumns?: boolean;
    columnWidth?: number;
    blocks: StatblockItem[];
}

export interface DefaultLayout extends Layout {
    edited?: boolean;
    removed?: boolean;
    version?: number;
    updatable?: boolean;
}

export type DiceParsing = {
    regex: string;
    parser: string;
    id: string;
    desc?: string;
};